+++ blueprints/app/index.js
@@ -21,8 +21,9 @@ module.exports = {
     let rawName = entity.name;
     let name = stringUtil.dasherize(rawName);
     let namespace = stringUtil.classify(rawName);
+    let embroider = isExperimentEnabled('EMBROIDER') || options.embroider;
 
-    let hasOptions = !options.welcome || options.yarn;
+    let hasOptions = !options.welcome || options.yarn || embroider;
     let blueprintOptions = '';
     if (hasOptions) {
       let indent = `\n            `;
@@ -30,7 +31,9 @@ module.exports = {
 
       blueprintOptions =
         indent +
-        [!options.welcome && '"--no-welcome"', options.yarn && '"--yarn"'].filter(Boolean).join(',\n            ') +
+        [!options.welcome && '"--no-welcome"', options.yarn && '"--yarn"', embroider && '"--embroider"']
+          .filter(Boolean)
+          .join(',\n            ') +
         outdent;
     }
 
@@ -43,7 +46,7 @@ module.exports = {
       welcome: options.welcome,
       blueprint: 'app',
       blueprintOptions,
-      embroider: isExperimentEnabled('EMBROIDER'),
+      embroider,
       lang: options.lang,
     };
   },