+++ tests/fixtures/app/with-blueprint-override-lint-fail/.prettierrc.js
@@ -1,0 +1,5 @@
+'use strict';
+
+module.exports = {
+  singleQuote: true, // blueprint overrides use double quotes
+};