+++ tests/fixtures/app/with-blueprint-override-lint-fail/.template-lintrc.js
@@ -1,0 +1,8 @@
+'use strict';
+
+module.exports = {
+  extends: 'recommended',
+  rules: {
+    'require-button-type': true, // blueprint override has missing button type
+  },
+};