+++ tests/fixtures/brocfile-tests/additional-trees/ember-cli-build.js
@@ -1,5 +1,6 @@
 const EmberApp = require('ember-cli/lib/broccoli/ember-app');
 const Funnel = require('broccoli-funnel');
+const { isExperimentEnabled } = require('ember-cli/lib/experiments');
 
 module.exports = function (defaults) {
   let app = new EmberApp(defaults, {});
@@ -8,6 +9,13 @@ module.exports = function (defaults) {
     srcDir: '/',
     destDir: '/assets'
   });
+
+  if (isExperimentEnabled('EMBROIDER')) {
+    const { Webpack } = require('@embroider/webpack');
+    return require('@embroider/compat').compatBuild(app, Webpack, {
+      extraPublicTrees: [funnel]
+    });
+  }
 
   return app.toTree(funnel);
 };