+++ tests/fixtures/brocfile-tests/app-import-custom-transform/ember-cli-build.js
@@ -1,5 +1,5 @@
 const EmberApp = require('ember-cli/lib/broccoli/ember-app');
-const map = require('broccoli-stew').map;
+const { isExperimentEnabled } = require('ember-cli/lib/experiments');
 
 module.exports = function (defaults) {
   var app = new EmberApp(defaults, {
@@ -13,6 +13,11 @@ module.exports = function (defaults) {
     ],
     outputFile: '/assets/output.js'
   });
+
+  if (isExperimentEnabled('EMBROIDER')) {
+    const { Webpack } = require('@embroider/webpack');
+    return require('@embroider/compat').compatBuild(app, Webpack);
+  }
 
   return app.toTree();
 };