+++ tests/fixtures/brocfile-tests/auto-run-false/ember-cli-build.js
@@ -1,9 +1,15 @@
 const EmberApp = require('ember-cli/lib/broccoli/ember-app');
+const { isExperimentEnabled } = require('ember-cli/lib/experiments');
 
 module.exports = function (defaults) {
   var app = new EmberApp(defaults, {
     autoRun: false
   });
+
+  if (isExperimentEnabled('EMBROIDER')) {
+    const { Webpack } = require('@embroider/webpack');
+    return require('@embroider/compat').compatBuild(app, Webpack);
+  }
 
   return app.toTree();
 };