+++ tests/fixtures/brocfile-tests/auto-run-true/config/targets.js
@@ -1,0 +1,10 @@
+const browsers = [
+  'last 1 Chrome versions',
+  'last 1 Firefox versions',
+  'last 1 Safari versions',
+];
+
+module.exports = {
+  browsers,
+  node: 'current',
+};