+++ tests/fixtures/brocfile-tests/custom-ember-env/config/environment.js
@@ -1,6 +1,7 @@
 module.exports = function() {
   return  {
     modulePrefix: 'some-cool-app',
+    rootURL: '/',
     EmberENV: {
       asdflkmawejf: ';jlnu3yr23'
     },