+++ tests/fixtures/brocfile-tests/custom-output-paths/ember-cli-build.js
@@ -1,4 +1,5 @@
 const EmberApp = require('ember-cli/lib/broccoli/ember-app');
+const { isExperimentEnabled } = require('ember-cli/lib/experiments');
 
 module.exports = function (defaults) {
   var app = new EmberApp(defaults, {
@@ -24,6 +25,11 @@ module.exports = function (defaults) {
       }
     }
   });
+
+  if (isExperimentEnabled('EMBROIDER')) {
+    const { Webpack } = require('@embroider/webpack');
+    return require('@embroider/compat').compatBuild(app, Webpack);
+  }
 
   return app.toTree();
 };