+++ tests/fixtures/help/with-addon-commands.js
@@ -256,6 +256,12 @@ module.exports = {
           required: false
         },
         {
+          name: 'lint-fix',
+          default: true,
+          key: 'lintFix',
+          required: false
+        },
+        {
           name: 'in',
           default: null,
           description: 'Runs a blueprint against an in repo addon. A path is expected, relative to the root of the project.',
@@ -404,6 +410,12 @@ module.exports = {
           required: false
         },
         {
+          name: 'lint-fix',
+          default: true,
+          key: 'lintFix',
+          required: false
+        },
+        {
           name: 'welcome',
           key: 'welcome',
           description: 'Installs and uses {{ember-welcome-page}}. Use --no-welcome to skip it.',
@@ -426,6 +438,13 @@ module.exports = {
           name: 'lang',
           key: 'lang',
           description: 'Sets the base human language of the application via index.html',
+          required: false
+        },
+        {
+          default: false,
+          key: 'embroider',
+          name: 'embroider',
+          description: 'Enables the build system to use Embroider',
           required: false
         }
       ],
@@ -537,6 +556,13 @@ module.exports = {
           name: 'lang',
           key: 'lang',
           description: 'Sets the base human language of the application via index.html',
+          required: false
+        },
+        {
+          default: false,
+          key: 'embroider',
+          name: 'embroider',
+          description: 'Enables the build system to use Embroider',
           required: false
         }
       ],