+++ tests/fixtures/preprocessor-tests/app-with-addon-with-preprocessors-3/app/components/should-not-be-preprocessed.js
@@ -1,5 +1,5 @@
 import Component from '@ember/component';
 
 export default Component.extend({
-  foo: __PREPROCESSOR_REPLACEMENT_TOKEN__
+  foo_in_app: __PREPROCESSOR_REPLACEMENT_TOKEN__
 });